Tidy up uses of LLVM headers
