+++ compiler/codegen/WhileDoStmt.cpp
@@ -23,6 +23,12 @@
 #include "build.h"
 #include "CForLoop.h"
 #include "codegen.h"
+#include "LayeredValueTable.h"
+
+#ifdef HAVE_LLVM
+#include "llvm/IR/Module.h"
+#endif
+
 
 /************************************ | *************************************
 *                                                                           *
