+++ compiler/codegen/symbol.cpp
@@ -28,6 +28,7 @@
 #include "AstVisitor.h"
 #include "astutil.h"
 #include "build.h"
+#include "clangUtil.h"
 #include "codegen.h"
 #include "CollapseBlocks.h"
 #include "docsDriver.h"
@@ -36,7 +37,9 @@
 #include "files.h"
 #include "intlimits.h"
 #include "iterator.h"
+#include "LayeredValueTable.h"
 #include "llvmDebug.h"
+#include "llvmUtil.h"
 #include "misc.h"
 #include "optimizations.h"
 #include "passes.h"
@@ -53,6 +56,14 @@
 
 #include <inttypes.h>
 #include <stdint.h>
+
+#ifdef HAVE_LLVM
+#include "llvm/IR/LegacyPassManager.h"
+#include "llvm/IR/Module.h"
+#include "llvm/IR/Verifier.h"
+
+#include "llvm/Support/raw_os_ostream.h"
+#endif
 
 /******************************** | *********************************
 *                                                                   *
