--- compiler/include/clangSupport.h
@@ -1,105 +1,0 @@
-/*
- * Copyright 2004-2017 Cray Inc.
- * Other additional copyright holders may be indicated within.
- * 
- * The entirety of this work is licensed under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * 
- * You may obtain a copy of the License at
- * 
- *     http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef _CLANGSUPPORT_H_
-#define _CLANGSUPPORT_H_
-
-#include "llvmUtil.h"
-
-#ifdef HAVE_LLVM
-
-// Workaround problem with 'exit' macro
-// 'exit' might be needed to work from LLVM headers.
-#ifdef exit
-#define CLANGSUPPORT_EXIT_WAS exit
-#undef exit
-#endif
-
-#include "llvm/ADT/StringRef.h"
-#include "llvm/ADT/SmallString.h"
-
-#include "clang/Basic/DiagnosticOptions.h"
-#include "clang/Lex/HeaderSearchOptions.h"
-#include "clang/Lex/PreprocessorOptions.h"
-
-#include "clang/Frontend/TextDiagnosticPrinter.h"
-#include "clang/AST/ASTContext.h"
-#include "clang/AST/TypeLoc.h"
-#include "clang/Basic/LangOptions.h"
-#include "clang/Basic/FileSystemOptions.h"
-#include "clang/Basic/SourceManager.h"
-#include "clang/Lex/HeaderSearch.h"
-#include "clang/Basic/FileManager.h"
-#include "clang/Frontend/Utils.h"
-#include "clang/Basic/Version.h"
-#include "clang/Basic/TargetOptions.h"
-#include "clang/Basic/TargetInfo.h"
-#include "clang/Lex/Preprocessor.h"
-#include "clang/Frontend/FrontendOptions.h"
-#include "clang/Basic/IdentifierTable.h"
-#include "clang/Basic/Builtins.h"
-#include "clang/AST/ASTContext.h"
-#include "clang/AST/ASTConsumer.h"
-#include "clang/Sema/Sema.h"
-#include "clang/AST/DeclBase.h"
-#include "clang/AST/Type.h"
-#include "clang/AST/Decl.h"
-#include "clang/Sema/Lookup.h"
-#include "clang/Sema/Ownership.h"
-#include "clang/AST/DeclGroup.h"
-#include "clang/Parse/Parser.h"
-#include "clang/Parse/ParseAST.h"
-#include "clang/CodeGen/ModuleBuilder.h"
-#include "clang/Frontend/CodeGenOptions.h"
-#include "clang/CodeGen/CodeGenAction.h"
-#include "clang/Driver/Compilation.h"
-#include "clang/Driver/Driver.h"
-#include "clang/Driver/Tool.h"
-#include "clang/Driver/Action.h"
-#include "clang/Frontend/CompilerInvocation.h"
-#include "clang/Frontend/CompilerInstance.h"
-#include "clang/Frontend/FrontendDiagnostic.h"
-#include "clang/Frontend/TextDiagnosticPrinter.h"
-#include "clang/Frontend/FrontendOptions.h"
-#include "clang/Frontend/Utils.h"
-#include "llvm/Support/FileSystem.h"
-#include "llvm/Support/Host.h"
-#include "llvm/Support/TargetSelect.h"
-#include "llvm/Support/Path.h"
-#include "llvm/Support/ToolOutputFile.h"
-
-#include "llvm/Bitcode/BitcodeReader.h"
-#include "llvm/Bitcode/BitcodeWriter.h"
-
-#include "llvm/IR/Verifier.h"
-
-
-#include "llvm/ADT/DenseMap.h"
-
-#include "llvm/Transforms/IPO/PassManagerBuilder.h"
-#include "llvm/Support/raw_ostream.h"
-
-#ifdef CLANGSUPPORT_EXIT_WAS
-#define exit CLANGSUPPORT_EXIT_WAS
-#undef CLANGSUPPORT_EXIT_WAS
-#endif
-
-#endif
-
-#endif
